Transition BatchStrategy to interface.
...and use long return type instead of int for cleanCache.
